attribute-sync-plugin: groups_to_roles does not verify ldapuser group membership correctly
The uniqueMember attribute of a groupOfUniqueNames could contain spaces
between the comma seperator, the distinguished name of a user could do
that too. If both ldap cn's are created equal the assignment for a
related wato roles will work. If one has any spaces after the seperator
(which is conform to RFC 4514 and RFC 1779) it will not.

As example we assume that:

UserDN:
  dn: cn=foo, ou=people, dc=bar, dc=de

GroupDN:
  dn: cn=group,ou=groups,dc=bar,dc=de

Ldapsearch for group content:
  ldapsearch -LLL -b "cn=group,ou=groups,dc=bar,dc=de" -h ... "(objectclass=groupOfUniqueNames)" uniqueMember
    dn: cn=group,ou=groups,dc=bar,dc=de
    ufn: group, groups, bar, de
    uniqueMember: cn=foo,ou=people,dc=bar,dc=de
    uniqueMember: cn=bar,ou=people,dc=bar,dc=de

Representation of the group ldap_groups[dn] at ldap.py line 1715:

  {
    u'cn=group,ou=foo,dc=bar,dc=de': {
      'cn': u'group',
      'members': ['cn=foo,ou=people,dc=bar,dc=de',
                  'cn=bar,ou=people,dc=bar,dc=de'
      ]
    }
  }

The user foo is created with spaces between the comma-seperator. The
condition "user_cmp_val in ldap_groups[dn]['members']" at line 1715-1716
of ldap.py will never match. The result is that the user will never be a
member of the expected and related wato role.